Fix integration

+++ lib/cadet/jobs/autograder/plagiarism_checker.ex
@@ -10,13 +10,18 @@ defmodule Cadet.Autograder.PlagiarismChecker do
 
   import Ecto.Query
 
-  @bucket_name "stg-cadet-plagiarism-reports" # TODO: change to env var
-  @plagiarism_script_path "../plag/TA_CS1101S/mosspy_submission.py" # TODO: change to env var
+  # TODO: change to env var
+  @bucket_name "stg-cadet-plagiarism-reports"
+  # TODO: change to env var
+  @plagiarism_script_path "../grader/TA_CS1101S/mosspy_submission.py"
 
   def perform(assessment_id) when is_ecto_id(assessment_id) do
     Logger.info("Running plagiarism check on Assessment #{assessment_id}")
 
-    System.cmd("python", [@plagiarism_script_path, "--assessment_id", to_string(assessment_id)])
+    script_result =
+      System.cmd("python", [@plagiarism_script_path, "--assessment_id", to_string(assessment_id)])
+
+    script_result
     |> elem(1)
     |> zip_results()
     |> store()
@@ -24,6 +29,7 @@ defmodule Cadet.Autograder.PlagiarismChecker do
 
   def store(assessment_id) when is_ecto_id(assessment_id) do
     file_name = "assessment_#{assessment_id}.zip"
+
     assessment_title =
       Cadet.Assessments.Assessment
       |> where(id: ^assessment_id)
@@ -36,12 +42,11 @@ defmodule Cadet.Autograder.PlagiarismChecker do
       |> ExAws.request!()
 
     # if response status < 400, the transaction was successful.
-    cond do
-      Map.get(response, :status_code) < 400 ->
-        File.rm(file_name)
-        File.rm_rf("submissions")
-      true ->
-        raise inspect(response)
+    if Map.get(response, :status_code) < 400 do
+      File.rm(file_name)
+      File.rm_rf("submissions")
+    else
+      raise inspect(response)
     end
   end
 
@@ -50,8 +55,9 @@ defmodule Cadet.Autograder.PlagiarismChecker do
       "-r",
       "assessment_#{assessment_id}.zip",
       "submissions/assessment#{assessment_id}/report/",
-      "submissions/assessment#{assessment_id}/assessment_report_#{assessment_id}.html",
+      "submissions/assessment#{assessment_id}/assessment_report_#{assessment_id}.html"
     ])
+
     assessment_id
   end
 end
